refactor: Move trace package into outer
