+++ advanced/src/main/java/hello/advanced/app/v1/OrderServiceV1.java
@@ -1,7 +1,7 @@
 package hello.advanced.app.v1;
 
-import hello.advanced.app.trace.TraceStatus;
-import hello.advanced.app.trace.hellotrace.HelloTraceV1;
+import hello.advanced.trace.TraceStatus;
+import hello.advanced.trace.hellotrace.HelloTraceV1;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
