+++ advanced/src/main/java/hello/advanced/app/v2/OrderControllerV2.java
@@ -1,7 +1,7 @@
 package hello.advanced.app.v2;
 
-import hello.advanced.app.trace.TraceStatus;
-import hello.advanced.app.trace.hellotrace.HelloTraceV2;
+import hello.advanced.trace.TraceStatus;
+import hello.advanced.trace.hellotrace.HelloTraceV2;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
