+++ advanced/src/main/java/hello/advanced/app/v2/OrderRepositoryV2.java
@@ -1,8 +1,8 @@
 package hello.advanced.app.v2;
 
-import hello.advanced.app.trace.TraceId;
-import hello.advanced.app.trace.TraceStatus;
-import hello.advanced.app.trace.hellotrace.HelloTraceV2;
+import hello.advanced.trace.TraceId;
+import hello.advanced.trace.TraceStatus;
+import hello.advanced.trace.hellotrace.HelloTraceV2;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
