+++ advanced/src/main/java/hello/advanced/trace/TraceId.java
@@ -1,4 +1,4 @@
-package hello.advanced.app.trace;
+package hello.advanced.trace;
 
 import java.util.UUID;
