+++ advanced/src/main/java/hello/advanced/trace/TraceStatus.java
@@ -1,4 +1,4 @@
-package hello.advanced.app.trace;
+package hello.advanced.trace;
 
 public class TraceStatus {
